openproject-widget/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
httpx==0.26.0
python-multipart==0.0.9
pydantic==2.6.1
python-dotenv==1.0.1
markdown==3.5.2

